pyproject.toml
[tool.poetry]
name = "modded-ftl-executor"
version = "0.1.0"
description = "Automatically provides modded environment for the game FTL."
authors = ["nedsociety <nedsociety@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
flet = "^0.1.54"
"fluent.runtime" = "^0.3.1"
requests = "^2.28.1"
javaproperties = "^0.8.1"
vdf = "^3.4"
dacite = "^1.6.0"
Babel = "^2.10.3"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
pyinstaller = "^5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
